Move Devices service to auth
