+++ libs/common/src/auth/abstractions/device-trust-crypto.service.abstraction.ts
@@ -1,6 +1,6 @@
-import { DeviceResponse } from "../../abstractions/devices/responses/device.response";
 import { EncString } from "../../platform/models/domain/enc-string";
 import { DeviceKey, UserKey } from "../../platform/models/domain/symmetric-crypto-key";
+import { DeviceResponse } from "../abstractions/devices/responses/device.response";
 
 export abstract class DeviceTrustCryptoServiceAbstraction {
   /**
