+++ libs/common/src/auth/abstractions/devices-api.service.abstraction.ts
@@ -1,5 +1,5 @@
-import { DeviceResponse } from "../../abstractions/devices/responses/device.response";
 import { ListResponse } from "../../models/response/list.response";
+import { DeviceResponse } from "../abstractions/devices/responses/device.response";
 import { SecretVerificationRequest } from "../models/request/secret-verification.request";
 import { UpdateDevicesTrustRequest } from "../models/request/update-devices-trust.request";
 import { ProtectedDeviceResponse } from "../models/response/protected-device.response";
