+++ libs/common/src/auth/abstractions/devices/devices.service.abstraction.ts
