+++ libs/common/src/auth/services/devices/devices.service.implementation.ts
@@ -1,10 +1,10 @@
 import { Observable, defer, map } from "rxjs";
 
+import { ListResponse } from "../../../models/response/list.response";
 import { DevicesServiceAbstraction } from "../../abstractions/devices/devices.service.abstraction";
 import { DeviceResponse } from "../../abstractions/devices/responses/device.response";
 import { DeviceView } from "../../abstractions/devices/views/device.view";
-import { DevicesApiServiceAbstraction } from "../../auth/abstractions/devices-api.service.abstraction";
-import { ListResponse } from "../../models/response/list.response";
+import { DevicesApiServiceAbstraction } from "../../abstractions/devices-api.service.abstraction";
 
 /**
  * @class DevicesServiceImplementation
